sending to master

+++ api/src/database/CourseRegistrationHelper.ts
@@ -1,13 +1,13 @@
 const HH = require("./HelperHelper")
 
-import {CourseRegistration, DBCourseRegistration} from '../../../models/CourseRegistration';
-import {localRoles} from '../../../enums/localRoleEnum'
+import {CourseRegistration, DBCourseRegistration, convert} from '../../../models/CourseRegistration';
+import {localRole, checkEnum} from '../../../enums/localRoleEnum'
 import RolePermissionHelper from './RolePermissionsHelper'
 /**
  * courseID, userID, role, permission
  * @Author Rens Leendertz
  */
-const pool = HH.pool
+const {pool, extract, map, one} = HH
 export default class CourseRegistrationHelper {
 
 	static toBin = RolePermissionHelper.toBin
@@ -15,17 +15,8 @@ export default class CourseRegistrationHelper {
 	/**
 	 * return all entries in this table, with permissions set correctly
 	 */
-	 static getAllEntries(){
-		return new Promise((
-				resolve : (result : CourseRegistration[]) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._getAllEntries(resolve, reject))
-	}
-
-	static _getAllEntries(
-			onSuccess : (result : CourseRegistration[]) => void, 
-			onFailure : (error : Error) => void) {
-		pool.query(`SELECT 
+	 static getAllEntries() {
+		return pool.query(`SELECT 
 				userID, 
 				courseID, 
 				courseRole, 
@@ -34,25 +25,15 @@ export default class CourseRegistrationHelper {
 							 WHERE courseRoleID=courseRole
 							) AS permission
 			FROM "CourseRegistration"`)
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(res.rows.map(CourseRegistrationHelper.DBToI)))
-		.catch(onFailure)
+		.then(extract).then(map(convert))
+		
 	}
 
 	/**
 	 * get all users entered in a specific course. permissions set correctly
 	 */
-	static getEntriesByCourse(courseID : string){
-		return new Promise((
-				resolve : (result : CourseRegistration[]) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._getEntriesByCourse(courseID, resolve,reject))
-	}
-
-	static _getEntriesByCourse(
-			courseID : string, 
-			onSuccess : (result : CourseRegistration[]) => void, 
-			onFailure : (error : Error) => void) {
-		pool.query(`SELECT 
+	static getEntriesByCourse(courseID : string) {
+		return pool.query(`SELECT 
 				userID, 
 				courseID, 
 				courseRole, 
@@ -62,25 +43,14 @@ export default class CourseRegistrationHelper {
 							) AS permission
 			FROM "CourseRegistration"
 			WHERE courseID=$1`, [courseID])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(res.rows.map(CourseRegistrationHelper.DBToI)))
-		.catch(onFailure)
+		.then(extract).then(map(convert))
 	}
 
 	/**
 	 * get all courses a user is entered into. permissions set correctly
 	 */
-	static getEntriesByUser(userID : string){
-		return new Promise((
-				resolve : (result : CourseRegistration[]) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._getEntriesByUser(userID, resolve,reject))
-	}
-
-	static _getEntriesByUser(
-			userID : string,
-			onSuccess : (result : CourseRegistration[]) => void, 
-			onFailure : (error : Error) => void) {
-		pool.query(`SELECT 
+	static getEntriesByUser(userID : string) {
+		return pool.query(`SELECT 
 				userID, 
 				courseID, 
 				courseRole, 
@@ -90,35 +60,24 @@ export default class CourseRegistrationHelper {
 							) AS permission
 			FROM "CourseRegistration" 
 			WHERE userID=$1`, [userID])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(res.rows.map(CourseRegistrationHelper.DBToI)))
-		.catch(onFailure)
+		.then(extract).then(map(convert))
+		
 	}
 
 	/**
 	 * add a new entry, all is required but permission. This defaults to no elevated permissions.
 	 */
 	static addEntry(entry : CourseRegistration){
-		return new Promise((
-				resolve : (result : CourseRegistration) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._addEntry(entry, resolve,reject))
-	}
-
-	static _addEntry(
-			entry : CourseRegistration,
-			onSuccess : (result : CourseRegistration) => void, 
-			onFailure : (error : Error) => void){
 		const {
-			courseid,
-			userid,
+			courseID,
+			userID,
 			role,
 			permission = 0
 		} = entry
-		pool.query(`INSERT INTO \"CourseRegistration\" 
+		return pool.query(`INSERT INTO \"CourseRegistration\" 
 			VALUES ($1,$2,$3,$4) 
-			RETURNING *`, [courseid, userid, role, CourseRegistrationHelper.toBin(permission)])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(CourseRegistrationHelper.DBToI(res.rows[0])))
-		.catch(onFailure)
+			RETURNING *`, [courseID, userID, role, CourseRegistrationHelper.toBin(permission)])
+		.then(extract).then(map(convert)).then(one)
 	}
 
 	/**
@@ -127,28 +86,17 @@ export default class CourseRegistrationHelper {
 	 * permission field will be ignored
 	 */
 	static updateRole(entry : CourseRegistration){
-		return new Promise((
-				resolve : (result : CourseRegistration) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._updateRole(entry, resolve,reject))
-	}
-
-	static _updateRole(
-			entry : CourseRegistration,
-			onSuccess : (result : CourseRegistration) => void,
-			onFailure : (error : Error) => void){
 		const {
-			courseid,
-			userid,
+			courseID,
+			userID,
 			role
 		} = entry
-		pool.query(`UPDATE \"CourseRegistration\" SET 
+		return pool.query(`UPDATE \"CourseRegistration\" SET 
 			courseRole=COALESCE($3, courseRole)
 			WHERE courseID=$1 AND userID=$2
 			RETURNING *
-			`, [courseid, userid, role])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(CourseRegistrationHelper.DBToI(res.rows[0])))
-		.catch(onFailure)
+			`, [courseID, userID, role])
+		.then(extract).then(map(convert)).then(one)
 	}
 
 	/**
@@ -157,28 +105,17 @@ export default class CourseRegistrationHelper {
 	 * the role field will be ignored, others are mandatory.
 	 */
 	static addPermission(entry : CourseRegistration){
-		return new Promise((
-				resolve : (result : CourseRegistration) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._addPermission(entry,resolve,reject))
-	}
-
-	static _addPermission(
-			entry : CourseRegistration,
-			onSuccess : (result : CourseRegistration) => void,
-			onFailure : (error : Error) => void){
 		const {
-			courseid,
-			userid,
+			courseID,
+			userID,
 			permission
 		} = entry
-		pool.query(`UPDATE \"CourseRegistration\" SET 
+		return pool.query(`UPDATE \"CourseRegistration\" SET 
 			permission=permission | $3
 			WHERE courseID=$1 AND userID=$2
 			RETURNING *
-			`, [courseid, userid, CourseRegistrationHelper.toBin(permission)])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(CourseRegistrationHelper.DBToI(res.rows[0])))
-		.catch(onFailure)
+			`, [courseID, userID, CourseRegistrationHelper.toBin(permission)])
+		.then(extract).then(map(convert)).then(one)
 	}
 
 	/**
@@ -188,63 +125,29 @@ export default class CourseRegistrationHelper {
 	 * the role field will be ignored, others are mandatory.
 	 */
 	static removePermission(entry : CourseRegistration){
-		return new Promise((
-				resolve : (result : CourseRegistration) => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._removePermission(entry, resolve,reject))
-	}
-
-	static _removePermission(
-			entry : CourseRegistration,
-			onSuccess : (result : CourseRegistration) => void,
-			onFailure : (error : Error) => void){
 		const {
-			courseid,
-			userid,
+			courseID,
+			userID,
 			permission
 		} = entry
-		pool.query(`UPDATE \"CourseRegistration\" SET 
+		return pool.query(`UPDATE \"CourseRegistration\" SET 
 			permission=permission & ~($3::bit(40))
 			WHERE courseID=$1 AND userID=$2
 			RETURNING *
-			`, [courseid, userid, CourseRegistrationHelper.toBin(permission)])
-		.then((res : {rows:DBCourseRegistration[]}) => onSuccess(CourseRegistrationHelper.DBToI(res.rows[0])))
-		.catch(onFailure)
+			`, [courseID, userID, CourseRegistrationHelper.toBin(permission)])
+		.then(extract).then(map(convert)).then(one)
 	}
 
 	/**
 	 * remove a user from a course. 
 	 * permission and role will be ignored.
 	 */
-	static deleteEntry(entry : CourseRegistration) : Promise<void>{
-		return new Promise((
-				resolve : () => void, 
-				reject: (error : Error) => void
-			) => CourseRegistrationHelper._deleteEntry(entry, resolve,reject))
-	}
-
-	static _deleteEntry(
-			entry : CourseRegistration,
-			onSuccess : () => void,
-			onFailure : (error : Error) => void){
+	static deleteEntry(entry : CourseRegistration){
 		const {
-			courseid,
-			userid
+			courseID,
+			userID
 		} = entry
-		pool.query("DELETE FROM \"CourseRegistration\" WHERE courseID=$1 AND userID=$2"
-			, [courseid, userid])
-		.then(onSuccess)
-		.catch(onFailure)
-
-	}
-
-	private static DBToI(entry : DBCourseRegistration) : CourseRegistration{
-		if (CourseRegistrationHelper.checkEnum(entry.courserole)){
-			return {...entry, role:localRoles[entry.courserole]}
-		}
-		throw new Error('non-existent enum type from db: '+entry.courserole)
-	}
-	private static checkEnum(role : string) : role is keyof typeof localRoles { 
-		return role in localRoles
+		return pool.query("DELETE FROM \"CourseRegistration\" WHERE courseID=$1 AND userID=$2 RETURNING *", [courseID, userID])
+		.then(extract).then(map(convert)).then(one)
 	}
 }
